tasks/Bank backend/requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cloudinary==1.30.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
pkg_resources==0.0.0
PyJWT==2.5.0
pyserial==3.5
python-qt==0.50
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
